Reverted to commit 529ff069e8cae5942cb7e86594e11a94e05e98fd

+++ src/components/admin/PropertySpecialPricingEnhanced.tsx
@@ -75,7 +75,7 @@ export const PropertySpecialPricingEnhanced = ({
 
     setLoading(true);
     try {
-      const price = parseFloat(specialPrice) || null;
+      const price = parseInt(specialPrice) || null;
 
       for (const date of selectedDates) {
         const { error } = await supabase.from("availability").upsert(
